Add Timeline page and navigation
