+++ package.json
@@ -16,9 +16,11 @@
     "@react-three/drei": "^10.1.2",
     "@react-three/fiber": "^9.1.2",
     "@react-three/postprocessing": "^3.0.4",
+    "@types/react-router-dom": "^5.3.3",
     "framer-motion": "^12.17.3",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "react-router-dom": "^7.6.2",
     "three": "^0.176.0",
     "three-stdlib": "^2.36.0"
   },
