+++ src/components/TopButtonBar.tsx
@@ -1,8 +1,10 @@
 import '../styles/TopButtonBar.css';
+import { Link } from 'react-router-dom';
 
 export default function TopButtonBar() {
   return (
     <div className="top-button-bar">
+      <Link className="glow-button" to="/timeline">Timeline</Link>
       <button className="glow-button">Projekte</button>
       <button className="glow-button">Über mich</button>
       <button className="glow-button">Kontakt</button>
