+++ src/data/events.ts
@@ -1,0 +1,22 @@
+import type { TimelineEvent } from '../components/Timeline';
+
+export const events: TimelineEvent[] = [
+  {
+    id: '1',
+    title: 'Geburt',
+    date: '1990',
+    description: 'Ich wurde geboren.',
+  },
+  {
+    id: '2',
+    title: 'Studium',
+    date: '2010',
+    description: 'Beginn meines Studiums.',
+  },
+  {
+    id: '3',
+    title: 'Erster Job',
+    date: '2015',
+    description: 'Start in der IT Branche.',
+  },
+];
